refactor: create file for model creation
