+++ src/shared/modules/offer/offer.entity.ts
@@ -1,6 +1,6 @@
-import { defaultClasses, getModelForClass, modelOptions, prop } from '@typegoose/typegoose';
-import { City, Offer } from '../../types/index.js';
 import { UserEntity } from '../user/user.entity.js';
+import { defaultClasses, modelOptions, prop } from '@typegoose/typegoose';
+import { City, Offer } from '../../types/index.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
 export interface OfferEntity extends defaultClasses.Base {}
@@ -95,7 +95,7 @@ export class OfferEntity extends defaultClasses.TimeStamps implements Offer {
 
   @prop({
     required: true,
-    ref: UserEntity,
+    ref: () => UserEntity,
   })
   public author!: UserEntity;
 
@@ -115,4 +115,3 @@ export class OfferEntity extends defaultClasses.TimeStamps implements Offer {
   public longitude!: number;
 }
 
-export const OfferModel = getModelForClass(OfferEntity);
