+++ src/shared/modules/user/user.container.ts
@@ -3,7 +3,8 @@ import { types } from '@typegoose/typegoose';
 import { UserService } from './user-service.interface.js';
 import { Component } from '../../types/index.js';
 import { DefaultUserService } from './default-user.service.js';
-import { UserEntity, UserModel } from './user.entity.js';
+import { UserEntity } from './user.entity.js';
+import { UserModel } from '../models.js';
 
 export function createUserContainer() {
   const userContainer = new Container();
